refactor(metrics): extract confusion-matrix base
- add `metrics/_confusion_matrix_metric.py` with `_ConfusionMatrixMetric`, holding the detection-matching and scoring machinery that `Precision`, `Recall` and `F1Score` each carried a copy of: `_compute`, `_compute_scores_for_classes`, `_compute_confusion_matrix`, `_detections_content`, `_make_empty_content`, `_filter_detections_by_size` and `_filter_predictions_and_targets_by_size`.
- drive the metric-specific parts through three hooks: the `_metric_name` class variable that error messages interpolate, the abstract `_score_from_confusion_matrix` formula, and the abstract `_build_result` factory that fills each result dataclass under its own field names.
- share result rendering through `_format_result`, `_result_to_pandas` and `_plot_result`, reading a `_ResultView` built by each result's `_result_view`; per-metric wording lives in `_ResultLabels`, and the score-column width is derived from the label rather than hard-coded per metric.
- keep `__init__`, `reset`, `update`, `compute`, `__str__`, `to_pandas` and `plot` defined on the public classes, since the API reference does not enable `inherited_members`.
- hoist the triplicated `(..., 3)` shape check into `_validate_confusion_matrix`, leaving `_compute_precision`, `_compute_recall` and `_compute_f1` as the formulas alone.
- drop `_match_detection_batch`, which all three metrics defined and none called, and the duplicated unreachable `raise ValueError` in `recall.py`.

---
